refactor: split 24 files exceeding 800-line limit into focused submodules
Split all files identified in #2020 that exceed the 800-line ceiling
from ARCHITECTURE.md crate boundary review criteria. 11 had already
been addressed by prior PRs; this handles the remaining 13:

Workspace crates (test extraction to submodule):
- aletheia/commands/memory.rs (961→787+173)
- episteme/serendipity.rs (947→557+389)
- hermeneus/complexity.rs (925→493+431)
- nous/competence.rs (832→538+293)
- dianoia/research.rs (867→388+478)

Workspace crates (logic extraction):
- theatron/tui/app/mod.rs (895→796+109 persistence.rs)
- organon/builtins/workspace_tests/path_ops.rs (870→489+409)
- organon/builtins/planning_tests.rs (803→606+208 mock)

Dead code removal:
- nous/pipeline/pipeline_tests.rs (897 lines, superseded by pipeline_tests/)

Desktop crate (excluded from workspace build):
- theatron/desktop/components/chat.rs (1026→529+496)
- theatron/desktop/views/meta/mod.rs (967→486+202+295)
- theatron/desktop/state/graph.rs (855→550+304)
- theatron/desktop/state/meta.rs (818→578+237)

Closes #2020

Gate-Passed: kanon 0.1.0
