+++ crates/organon/src/builtins/workspace_tests/mod.rs
@@ -1,3 +1,4 @@
 //! Tests for the workspace tool executor.
 mod filesystem_ops;
 mod path_ops;
+mod path_ops_navigation;
