+++ crates/theatron/desktop/src/views/meta/mod.rs
@@ -1,12 +1,12 @@
 //! Meta-insights view: the system looking at itself.
 
+mod assembly;
+mod charts;
 mod health;
 mod knowledge;
 mod performance;
 mod quality;
 mod reflection;
-
-use std::collections::HashMap;
 
 use dioxus::prelude::*;
 
@@ -18,6 +18,9 @@ use crate::state::meta::{
     SystemReflectionStore,
 };
 
+pub(crate) use charts::{BarChart, LineChart};
+
+use assembly::assemble_meta_data;
 use health::MemoryHealthSection;
 use knowledge::KnowledgeGrowthSection;
 use performance::AgentPerformanceSection;
@@ -376,205 +379,6 @@ fn AccordionSection(
     }
 }
 
-// -- SVG chart helpers --------------------------------------------------------
-
-/// Render a simple SVG line chart.
-#[component]
-pub(crate) fn LineChart(
-    data: Vec<crate::state::meta::DataPoint>,
-    width: f64,
-    height: f64,
-    color: &'static str,
-    show_labels: bool,
-) -> Element {
-    if data.is_empty() {
-        return rsx! {
-            div { style: "{MUTED_TEXT}", "No data available" }
-        };
-    }
-
-    let max_val = data
-        .iter()
-        .map(|p| p.value)
-        .fold(f64::NEG_INFINITY, f64::max)
-        .max(1.0);
-
-    let padding = 30.0;
-    let chart_w = width - padding * 2.0;
-    let chart_h = height - padding * 2.0;
-
-    let points: Vec<(f64, f64)> = data
-        .iter()
-        .enumerate()
-        .map(|(i, p)| {
-            let x = if data.len() > 1 {
-                padding + (i as f64 / (data.len() - 1) as f64) * chart_w
-            } else {
-                padding + chart_w / 2.0
-            };
-            let y = padding + chart_h - (p.value / max_val) * chart_h;
-            (x, y)
-        })
-        .collect();
-
-    let path = points
-        .iter()
-        .enumerate()
-        .map(|(i, (x, y))| {
-            if i == 0 {
-                format!("M {x:.1} {y:.1}")
-            } else {
-                format!("L {x:.1} {y:.1}")
-            }
-        })
-        .collect::<Vec<_>>()
-        .join(" ");
-
-    let viewbox = format!("0 0 {width} {height}");
-
-    rsx! {
-        svg {
-            width: "{width}",
-            height: "{height}",
-            view_box: "{viewbox}",
-            style: "display: block;",
-
-            // NOTE: Grid lines.
-            for i in 0..5u32 {
-                {
-                    let y = padding + (f64::from(i) / 4.0) * chart_h;
-                    rsx! {
-                        line {
-                            x1: "{padding}",
-                            y1: "{y:.1}",
-                            x2: "{padding + chart_w:.1}",
-                            y2: "{y:.1}",
-                            stroke: "#2a2a3a",
-                            stroke_width: "1",
-                        }
-                    }
-                }
-            }
-
-            path {
-                d: "{path}",
-                fill: "none",
-                stroke: "{color}",
-                stroke_width: "2",
-            }
-
-            for (x , y) in &points {
-                circle {
-                    cx: "{x:.1}",
-                    cy: "{y:.1}",
-                    r: "3",
-                    fill: "{color}",
-                }
-            }
-
-            if show_labels {
-                // NOTE: Show first and last labels only to avoid clutter.
-                if let Some(first) = data.first() {
-                    {
-                        let (x, _) = points[0];
-                        rsx! {
-                            text {
-                                x: "{x:.1}",
-                                y: "{height - 4.0:.1}",
-                                fill: "#666",
-                                font_size: "10",
-                                text_anchor: "start",
-                                "{first.label}"
-                            }
-                        }
-                    }
-                }
-                if data.len() > 1 {
-                    if let Some(last) = data.last() {
-                        {
-                            let (x, _) = points[points.len() - 1];
-                            rsx! {
-                                text {
-                                    x: "{x:.1}",
-                                    y: "{height - 4.0:.1}",
-                                    fill: "#666",
-                                    font_size: "10",
-                                    text_anchor: "end",
-                                    "{last.label}"
-                                }
-                            }
-                        }
-                    }
-                }
-            }
-        }
-    }
-}
-
-/// Render a horizontal bar chart.
-#[component]
-pub(crate) fn BarChart(
-    data: Vec<crate::state::meta::DataPoint>,
-    width: f64,
-    height: f64,
-    color: &'static str,
-) -> Element {
-    if data.is_empty() {
-        return rsx! {
-            div { style: "{MUTED_TEXT}", "No data available" }
-        };
-    }
-
-    let max_val = data
-        .iter()
-        .map(|p| p.value)
-        .fold(f64::NEG_INFINITY, f64::max)
-        .max(1.0);
-
-    let padding = 30.0;
-    let chart_w = width - padding * 2.0;
-    let chart_h = height - padding * 2.0;
-    let bar_width = (chart_w / data.len() as f64) * 0.7;
-    let gap = (chart_w / data.len() as f64) * 0.3;
-
-    let viewbox = format!("0 0 {width} {height}");
-
-    rsx! {
-        svg {
-            width: "{width}",
-            height: "{height}",
-            view_box: "{viewbox}",
-            style: "display: block;",
-
-            for (i , point) in data.iter().enumerate() {
-                {
-                    let bar_h = (point.value / max_val) * chart_h;
-                    let x = padding + (i as f64 * (bar_width + gap)) + gap / 2.0;
-                    let y = padding + chart_h - bar_h;
-                    rsx! {
-                        rect {
-                            x: "{x:.1}",
-                            y: "{y:.1}",
-                            width: "{bar_width:.1}",
-                            height: "{bar_h:.1}",
-                            fill: "{color}",
-                            rx: "2",
-                        }
-                        text {
-                            x: "{x + bar_width / 2.0:.1}",
-                            y: "{height - 4.0:.1}",
-                            fill: "#666",
-                            font_size: "9",
-                            text_anchor: "middle",
-                            "{point.label}"
-                        }
-                    }
-                }
-            }
-        }
-    }
-}
-
 // -- Data fetch ---------------------------------------------------------------
 
 async fn fetch_meta_data(cfg: &ConnectionConfig) -> FetchState<MetaData> {
@@ -680,288 +484,3 @@ async fn fetch_meta_data(cfg: &ConnectionConfig) -> FetchState<MetaData> {
     FetchState::Loaded(data)
 }
 
-/// Assemble all fetched data into the composite `MetaData` structure.
-#[expect(clippy::cast_precision_loss, reason = "display-only metrics")]
-fn assemble_meta_data(
-    health: HealthApiResponse,
-    metrics: MetricsApiResponse,
-    facts: Vec<FactEntry>,
-    entities: Vec<EntityEntry>,
-    timeline: Vec<TimelineEntry>,
-    sessions: Vec<SessionEntry>,
-    agents: Vec<AgentEntry>,
-) -> MetaData {
-    // -- Performance --
-    let mut scorecards = Vec::new();
-    for agent in &agents {
-        let agent_sessions: Vec<&SessionEntry> =
-            sessions.iter().filter(|s| s.nous_id == agent.id).collect();
-        let session_count = agent_sessions.len().max(1) as f64;
-        let total_messages: u32 = agent_sessions.iter().map(|s| s.message_count).sum();
-
-        scorecards.push(crate::state::meta::AgentScorecard {
-            agent_id: agent.id.clone(),
-            agent_name: if agent.name.is_empty() {
-                agent.id.clone()
-            } else {
-                agent.name.clone()
-            },
-            avg_tokens_per_response: 0.0,
-            tool_calls_per_session: 0.0,
-            tool_success_rate: 0.85,
-            distillation_frequency: 0.0,
-            avg_context_before_distill: 0.0,
-            messages_per_session: total_messages as f64 / session_count,
-            sessions_per_day: 0.0,
-            errors_per_session: 0.0,
-        });
-    }
-
-    let performance = AgentPerformanceStore {
-        scorecards,
-        anomalies: Vec::new(),
-        tokens_per_response_series: HashMap::new(),
-    };
-
-    // -- Quality --
-    let mut depth = crate::state::meta::DepthDistribution::default();
-    for s in &sessions {
-        match crate::state::meta::DepthDistribution::classify(s.message_count) {
-            "short" => depth.short += 1,
-            "medium" => depth.medium += 1,
-            _ => depth.long += 1,
-        }
-    }
-
-    let quality = QualityStore {
-        avg_turn_length: Vec::new(),
-        response_to_question_ratio: Vec::new(),
-        tool_call_density: Vec::new(),
-        thinking_time_ratio: Vec::new(),
-        depth_distribution: depth,
-        top_topics: Vec::new(),
-    };
-
-    // -- Knowledge growth --
-    let total_entity_count = entities.len() as u64;
-    let total_relationship_count: u32 = entities.iter().map(|e| e.relationship_count).sum();
-
-    let cumulative_entities: Vec<crate::state::meta::DataPoint> = timeline
-        .iter()
-        .scan(0u32, |acc, entry| {
-            *acc += entry.count;
-            Some(crate::state::meta::DataPoint {
-                label: entry.date.clone(),
-                value: f64::from(*acc),
-            })
-        })
-        .collect();
-
-    let new_per_period: Vec<crate::state::meta::DataPoint> = timeline
-        .iter()
-        .map(|e| crate::state::meta::DataPoint {
-            label: e.date.clone(),
-            value: f64::from(e.count),
-        })
-        .collect();
-
-    let cumulative_values: Vec<f64> = cumulative_entities.iter().map(|p| p.value).collect();
-    let acceleration = crate::state::meta::compute_acceleration(&cumulative_values);
-
-    // WHY: Entity type distribution from entity list.
-    let mut type_counts: HashMap<&str, u32> = HashMap::new();
-    for e in &entities {
-        let t = if e.entity_type.is_empty() {
-            "unknown"
-        } else {
-            e.entity_type.as_str()
-        };
-        *type_counts.entry(t).or_default() += 1;
-    }
-    let mut type_slices: Vec<crate::state::meta::EntityTypeSlice> = type_counts
-        .into_iter()
-        .enumerate()
-        .map(|(i, (t, count)): (usize, (&str, u32))| crate::state::meta::EntityTypeSlice {
-            entity_type: t.to_string(),
-            count,
-            color: crate::state::meta::ENTITY_TYPE_COLORS
-                [i % crate::state::meta::ENTITY_TYPE_COLORS.len()],
-        })
-        .collect();
-    type_slices.sort_by(|a, b| b.count.cmp(&a.count));
-
-    let current_entity_rate = new_per_period.last().map_or(0.0, |p| p.value);
-
-    let density_over_time = if total_entity_count > 0 {
-        vec![crate::state::meta::DataPoint {
-            label: "now".to_string(),
-            value: total_relationship_count as f64 / total_entity_count as f64,
-        }]
-    } else {
-        Vec::new()
-    };
-
-    let knowledge = KnowledgeGrowthStore {
-        total_entities: cumulative_entities,
-        new_entities_per_period: new_per_period,
-        total_relationships: Vec::new(),
-        new_relationships_per_period: Vec::new(),
-        density_over_time,
-        entity_type_distribution: type_slices,
-        current_entity_rate,
-        current_relationship_rate: 0.0,
-        acceleration,
-    };
-
-    // -- Memory health --
-    let avg_confidence = if facts.is_empty() {
-        0.0
-    } else {
-        facts.iter().map(|f| f.confidence).sum::<f64>() / facts.len() as f64
-    };
-
-    let orphan_count = entities
-        .iter()
-        .filter(|e| e.relationship_count <= 1)
-        .count();
-    let orphan_ratio = if entities.is_empty() {
-        0.0
-    } else {
-        orphan_count as f64 / entities.len() as f64
-    };
-
-    // WHY: Approximate staleness from facts with empty updated_at or old dates.
-    let stale_count = facts.iter().filter(|f| f.updated_at.is_empty()).count();
-    let staleness_ratio = if facts.is_empty() {
-        0.0
-    } else {
-        stale_count as f64 / facts.len() as f64
-    };
-
-    let health_score =
-        crate::state::meta::compute_health_score(avg_confidence, orphan_ratio, staleness_ratio);
-
-    // WHY: Build confidence distribution histogram (10 buckets of 0.1 width).
-    let mut confidence_buckets = vec![0u32; 10];
-    for f in &facts {
-        let idx = ((f.confidence * 10.0).floor() as usize).min(9);
-        confidence_buckets[idx] += 1;
-    }
-    let confidence_distribution: Vec<crate::state::meta::ConfidenceBucket> = confidence_buckets
-        .into_iter()
-        .enumerate()
-        .map(|(i, count)| {
-            let lo = i as f64 * 0.1;
-            let hi = lo + 0.1;
-            crate::state::meta::ConfidenceBucket {
-                range_label: format!("{lo:.1}-{hi:.1}"),
-                count,
-            }
-        })
-        .collect();
-
-    let recommendations = crate::state::meta::generate_recommendations(
-        staleness_ratio,
-        orphan_ratio,
-        avg_confidence,
-        current_entity_rate,
-    );
-
-    let mem_health = MemoryHealthStore {
-        health_score,
-        confidence_distribution,
-        stale_entities: Vec::new(),
-        orphan_ratio,
-        decay_pressure_count: 0,
-        health_over_time: vec![crate::state::meta::DataPoint {
-            label: "now".to_string(),
-            value: health_score,
-        }],
-        recommendations,
-    };
-
-    // -- Reflection --
-    let total_tokens = metrics.total_input_tokens + metrics.total_output_tokens;
-    let overview = crate::state::meta::SystemOverview {
-        uptime_seconds: health.uptime_seconds,
-        total_sessions: metrics.total_sessions,
-        total_tokens,
-        total_entities: total_entity_count,
-        total_cost_usd: metrics.total_cost_usd,
-    };
-
-    let efficiency = crate::state::meta::EfficiencyMetrics {
-        cost_per_entity: crate::state::meta::cost_per_entity(
-            metrics.total_cost_usd,
-            total_entity_count,
-        ),
-        cost_per_session: if metrics.total_sessions > 0 {
-            metrics.total_cost_usd / metrics.total_sessions as f64
-        } else {
-            0.0
-        },
-        tokens_per_entity: crate::state::meta::tokens_per_entity(total_tokens, total_entity_count),
-        cost_per_entity_trend: crate::state::meta::TrendDirection::Flat,
-    };
-
-    // WHY: Build heatmap from session created_at timestamps.
-    // Parse "YYYY-MM-DDTHH:MM:SS" → (day_of_week, hour).
-    let timestamps: Vec<(u8, u8)> = sessions
-        .iter()
-        .filter_map(|s| parse_timestamp_to_day_hour(&s.created_at))
-        .collect();
-    let heatmap = crate::state::meta::build_heatmap(&timestamps);
-
-    let reflection = SystemReflectionStore {
-        overview,
-        heatmap,
-        efficiency,
-        journal: Vec::new(),
-    };
-
-    MetaData {
-        performance,
-        quality,
-        knowledge,
-        health: mem_health,
-        reflection,
-    }
-}
-
-/// Extract (day_of_week, hour) from an ISO 8601 timestamp string.
-///
-/// Uses a basic parser -- no external date library dependency needed.
-fn parse_timestamp_to_day_hour(ts: &str) -> Option<(u8, u8)> {
-    // WHY: Minimal parsing for "YYYY-MM-DDTHH:..." format.
-    if ts.len() < 13 {
-        return None;
-    }
-    let hour: u8 = ts.get(11..13)?.parse().ok()?;
-    if hour >= 24 {
-        return None;
-    }
-
-    // WHY: Approximate day-of-week using Tomohiko Sakamoto's algorithm.
-    let year: i32 = ts.get(0..4)?.parse().ok()?;
-    let month: u32 = ts.get(5..7)?.parse().ok()?;
-    let day: u32 = ts.get(8..10)?.parse().ok()?;
-
-    if !(1..=12).contains(&month) || !(1..=31).contains(&day) {
-        return None;
-    }
-
-    let dow = day_of_week(year, month, day);
-    Some((dow, hour))
-}
-
-/// Tomohiko Sakamoto's day-of-week algorithm. Returns 0=Mon..6=Sun.
-fn day_of_week(mut year: i32, month: u32, day: u32) -> u8 {
-    const T: [i32; 12] = [0, 3, 2, 5, 0, 3, 5, 1, 4, 6, 2, 4];
-    if month < 3 {
-        year -= 1;
-    }
-    let dow =
-        (year + year / 4 - year / 100 + year / 400 + T[month as usize - 1] + day as i32) % 7;
-    // WHY: Sakamoto returns 0=Sun, convert to 0=Mon.
-    ((dow + 6) % 7) as u8
-}
